renaming Profile folders to Account
