+++ test/components/Account/AccountInputField.test.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { render } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
-import { AccountInputField } from '@components/Profile';
+import { AccountInputField } from '@components/Account';
 import '@testing-library/jest-dom/extend-expect';
 
 const MockProfileComponent = ({ name }) => {
